Title and text
Added the code for the title, also created another function for the text, for right now it is only used in the title but it should be useful for any place we have regular text.
It checks if the current token is not a special character, and if it is not then recursively calls back to itself until we get a special character, then it just stops and continues the flow of whatever function it is in.